fix: fix db error and avoid duplicate cids
Moves db error into the client to make it reusable
Avoids duplicate cidv0/v1 in the push pipeline for content and pins
