+++ packages/api/src/utils/db-client.js
@@ -1,5 +1,4 @@
 import { PostgrestClient, PostgrestQueryBuilder } from '@supabase/postgrest-js'
-import { DBError } from '../errors.js'
 
 /**
  * @typedef {import('./db-types').definitions} definitions
@@ -371,3 +370,20 @@ export class DBClient {
     }
   }
 }
+
+export class DBError extends Error {
+  /**
+   * @param {{
+   *   message: string
+   *   details: string
+   *   hint: string
+   *   code: string
+   * }} cause
+   */
+  constructor({ message, details, hint, code }) {
+    super(`${message}, details: ${details}, hint: ${hint}, code: ${code}`)
+    this.name = 'DBError'
+    this.code = DBError.CODE
+  }
+}
+DBError.CODE = 'ERROR_DB'
